requirements.txt
audioread==3.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
fbpca==1.0
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.43.1
ftfy==6.1.1
future==0.18.3
glfw==2.6.2
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
imgui==1.4.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lpips==0.1.4
MarkupSafe==2.1.2
matplotlib==3.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
ndi-python==5.1.1.1
networkx==3.0
numba==0.58.1
numpy==1.24.1
opencv-python==4.8.1.78
opensimplex==0.4.5
packaging==23.2
pandas==2.1.1
Pillow==9.3.0
platformdirs==3.11.0
pooch==1.7.0
proglog==0.1.10
psutil==5.9.6
PyAudio==0.2.13
pycparser==2.21
PyOpenGL==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-osc==1.8.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
# torch==2.0.1+cu117
# torchaudio==2.0.2+cu117
# torchvision==0.15.2+cu117
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
visualizer==0.0.10
wcwidth==0.2.8
wordcloud==1.9.2
timm==0.5.4
ninja==1.11.1.1
kmeans-pytorch==0.3
